refactor(plugins): move common fixtures into `consume/hive_simulators/conftest.py`
